started again. fixed rendering and delete/like buttons

+++ src/utils/api.js
@@ -25,19 +25,33 @@ class Api {
     }).then(this._checkResponse);
   }
 
-  putLike(id) {
-    return fetch(`${this._url}/cards/${id}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    }).then(this._checkResponse);
+  changeLikeCardStatus(id, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._url}/cards/like/${id}`, {
+        method: "PUT",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    } else {
+      return fetch(`${this._Url}/cards/like/${id}`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    }
   }
 
-  deleteLike(id) {
-    return fetch(`${this._url}/cards/${id}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    }).then(this._checkResponse);
-  }
+  // putLike(id) {
+  //   return fetch(`${this._url}/cards/${id}/likes`, {
+  //     method: "PUT",
+  //     headers: this._headers,
+  //   }).then(this._checkResponse);
+  // }
+
+  // deleteLike(id) {
+  //   return fetch(`${this._url}/cards/${id}/likes`, {
+  //     method: "DELETE",
+  //     headers: this._headers,
+  //   }).then(this._checkResponse);
+  // }
 
   deleteCard(id) {
     return fetch(`${this._url}/cards/${id}`, {
